@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
@@ -12,16 +13,20 @@ namespace Zaabee.NewtonsoftJson
         public static byte[] ToBytes(this object obj, JsonSerializerSettings settings = null) =>
             NewtonsoftJsonHelper.Serialize(obj, settings);
 
-        public static MemoryStream ToStream(this object obj, JsonSerializerSettings settings = null) =>
-            NewtonsoftJsonHelper.Pack(obj, settings);
+        public static MemoryStream ToStream(this object obj, JsonSerializerSettings settings = null,
+            Encoding encoding = null) =>
+            NewtonsoftJsonHelper.Pack(obj, settings, encoding);
 
-        public static async Task<MemoryStream> ToStreamAsync(this object obj, JsonSerializerSettings settings = null) =>
-            await NewtonsoftJsonHelper.PackAsync(obj, settings);
+        public static async Task<MemoryStream> ToStreamAsync(this object obj, JsonSerializerSettings settings = null,
+            Encoding encoding = null) =>
+            await NewtonsoftJsonHelper.PackAsync(obj, settings, encoding);
 
-        public static void PackTo(this object obj, Stream stream, JsonSerializerSettings settings = null) =>
-            NewtonsoftJsonHelper.Pack(obj, stream, settings);
+        public static void PackTo(this object obj, Stream stream, JsonSerializerSettings settings = null,
+            Encoding encoding = null) =>
+            NewtonsoftJsonHelper.Pack(obj, stream, settings, encoding);
 
-        public static async Task PackToAsync(this object obj, Stream stream, JsonSerializerSettings settings = null) =>
-            await NewtonsoftJsonHelper.PackAsync(obj, stream, settings);
+        public static async Task PackToAsync(this object obj, Stream stream, JsonSerializerSettings settings = null,
+            Encoding encoding = null) =>
+            await NewtonsoftJsonHelper.PackAsync(obj, stream, settings, encoding);
     }
 }
