@@ -5,10 +5,10 @@ namespace Zaabee.Xml
 {
     public static class TextReaderExtension
     {
-        public static T FromXml<T>(this TextReader textReader) =>
+        public static T ReadXml<T>(this TextReader textReader) =>
             XmlHelper.Deserialize<T>(textReader);
 
-        public static object FromXml(this TextReader textReader, Type type) =>
+        public static object ReadXml(this TextReader textReader, Type type) =>
             XmlHelper.Deserialize(type, textReader);
     }
 }
