@@ -5,10 +5,10 @@ namespace Zaabee.Xml
 {
     public static class XmlReaderExtension
     {
-        public static T FromXml<T>(this XmlReader xmlReader) =>
+        public static T ReadXml<T>(this XmlReader xmlReader) =>
             XmlHelper.Deserialize<T>(xmlReader);
 
-        public static object FromXml(this XmlReader xmlReader, Type type) =>
+        public static object ReadXml(this XmlReader xmlReader, Type type) =>
             XmlHelper.Deserialize(type, xmlReader);
     }
 }
