@@ -2,7 +2,6 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Xunit;
-using Zaabee.Utf8Json;
 
 namespace Zaabee.Utf8Json.UnitTest
 {
@@ -15,11 +14,12 @@ namespace Zaabee.Utf8Json.UnitTest
             var testModel = GetTestModel();
             var stream0 = new FileStream(".\\StreamTest0",FileMode.Create);
             await testModel.PackToAsync(stream0);
-            var stream1 = new FileStream(".\\StreamTest1",FileMode.Create);
-            await stream1.PackByAsync(testModel);
+            var stream1 = await testModel.ToStreamAsync();
+            var stream2 = await testModel.ToStreamAsync(typeof(TestModel));
 
             var unPackResult0 = await stream0.UnpackAsync<TestModel>();
             var unPackResult1 = await stream1.UnpackAsync<TestModel>();
+            var unPackResult2 = await stream2.UnpackAsync<TestModel>();
 
             Assert.Equal(
                 Tuple.Create(testModel.Id, testModel.Age, testModel.CreateTime, testModel.Name, testModel.Gender),
@@ -29,6 +29,10 @@ namespace Zaabee.Utf8Json.UnitTest
                 Tuple.Create(testModel.Id, testModel.Age, testModel.CreateTime, testModel.Name, testModel.Gender),
                 Tuple.Create(unPackResult1.Id, unPackResult1.Age, unPackResult1.CreateTime, unPackResult1.Name,
                     unPackResult1.Gender));
+            Assert.Equal(
+                Tuple.Create(testModel.Id, testModel.Age, testModel.CreateTime, testModel.Name, testModel.Gender),
+                Tuple.Create(unPackResult2.Id, unPackResult2.Age, unPackResult2.CreateTime, unPackResult2.Name,
+                    unPackResult2.Gender));
         }
 
         [Fact]
@@ -39,11 +43,12 @@ namespace Zaabee.Utf8Json.UnitTest
 
             var stream0 = new FileStream(".\\StreamNonGenericTest0",FileMode.Create);
             await testModel.PackToAsync(type, stream0);
-            var stream1 = new FileStream(".\\StreamNonGenericTest1",FileMode.Create);
-            await stream1.PackByAsync(type, testModel);
+            var stream1 = await testModel.ToStreamAsync();
+            var stream2 = await testModel.ToStreamAsync(typeof(TestModel));
 
             var unPackResult0 = (TestModel) await stream0.UnpackAsync(type);
             var unPackResult1 = (TestModel) await stream1.UnpackAsync(type);
+            var unPackResult2 = (TestModel) await stream2.UnpackAsync(type);
 
             Assert.Equal(
                 Tuple.Create(((TestModel) testModel).Id, ((TestModel) testModel).Age,
@@ -55,6 +60,11 @@ namespace Zaabee.Utf8Json.UnitTest
                     ((TestModel) testModel).CreateTime, ((TestModel) testModel).Name, ((TestModel) testModel).Gender),
                 Tuple.Create(unPackResult1.Id, unPackResult1.Age, unPackResult1.CreateTime, unPackResult1.Name,
                     unPackResult1.Gender));
+            Assert.Equal(
+                Tuple.Create(((TestModel) testModel).Id, ((TestModel) testModel).Age,
+                    ((TestModel) testModel).CreateTime, ((TestModel) testModel).Name, ((TestModel) testModel).Gender),
+                Tuple.Create(unPackResult2.Id, unPackResult2.Age, unPackResult2.CreateTime, unPackResult2.Name,
+                    unPackResult2.Gender));
         }
 
         private static TestModel GetTestModel()
